+++ kernels/common/scene_subdiv_mesh.cpp
@@ -106,7 +106,7 @@ namespace embree
         vertex_buffer_tags.resize(bid+1);
       }
       vertices[bid].newBuffer(device,size,stride);
-      setNumTimeSteps(vertices.size());
+      setNumTimeSteps((unsigned int)vertices.size());
       return vertices[bid].get();
     }
     else if (type >= RTC_USER_VERTEX_BUFFER0 && type < RTC_USER_VERTEX_BUFFER0+RTC_MAX_USER_VERTEX_BUFFERS)
@@ -198,7 +198,7 @@ namespace embree
         vertices.pop_back();
         vertex_buffer_tags.pop_back();
         }*/
-      setNumTimeSteps(vertices.size());
+      setNumTimeSteps((unsigned int)vertices.size());
     }
     else if (type >= RTC_USER_VERTEX_BUFFER0 && type < RTC_USER_VERTEX_BUFFER_(RTC_MAX_USER_VERTEX_BUFFERS))
     {
